use intptr_t where pertinent

+++ vm/interpreter/floats.cpp
@@ -93,9 +93,9 @@ void Floats::generate_tst(MacroAssembler* masm, Assembler::Condition cc) {
   masm->fnstsw_ax();
   masm->fpop();		// explicitly pop argument
   masm->testl(eax, mask);
-  masm->movl(eax, Address((int)&trueObj, relocInfo::external_word_type));
+  masm->movl(eax, Address((intptr_t)&trueObj, relocInfo::external_word_type));
   masm->jcc(cond, L);
-  masm->movl(eax, Address((int)&falseObj, relocInfo::external_word_type));
+  masm->movl(eax, Address((intptr_t)&falseObj, relocInfo::external_word_type));
   masm->bind(L);
 }
 
@@ -109,9 +109,9 @@ void Floats::generate_cmp(MacroAssembler* masm, Assembler::Condition cc) {
   masm->fwait();
   masm->fnstsw_ax();
   masm->testl(eax, mask);
-  masm->movl(eax, Address((int)&trueObj, relocInfo::external_word_type));
+  masm->movl(eax, Address((intptr_t)&trueObj, relocInfo::external_word_type));
   masm->jcc(cond, L);
-  masm->movl(eax, Address((int)&falseObj, relocInfo::external_word_type));
+  masm->movl(eax, Address((intptr_t)&falseObj, relocInfo::external_word_type));
   masm->bind(L);
 }
